Software flattened, over to docs.
