+++ docs/pluggable_app/index.md
@@ -36,6 +36,7 @@ protocols
 extend
 replace
 vary
-simple_predicates
+noqual
 scope
+context
 ```
