+++ docs/pluggable_app/replace.md
@@ -80,7 +80,7 @@ It shouldn't have to do the extra work to go find the implementation it wants to
 Fortunately Antidote has a better approach, [using `by_default`](https://antidote.readthedocs.io/en/stable/recipes/interface.html#default).
 We'll change the `Salutation` plugin to indicate that `DefaultSalutation` should _only_ be used if nothing else is provided.
 
-```{literalinclude} ../../src/antidote_book/replace/megastore_plugins/salutation/__init__.py
+```{literalinclude} ../../src/antidote_book/replace/megastore_plugins/salutation.py
 ---
 start-at: "@implements(Salutation)"
 end-at: class DefaultSalutation
