+++ src/antidote_book/context/megastore_plugins/greeting.py
@@ -9,11 +9,12 @@ from antidote import inject
 from antidote import injectable
 from antidote import interface
 
-from ..config import MegaStoreConfig
-from ..customer import Customer
-from ..customer import DefaultCustomer
-from ..customer import FrenchCustomer
-from ..greeter import Greeter
+from ..megastore.predicates import NotQualified
+from .config import MegaStoreConfig
+from .customer import Customer
+from .customer import FrenchCustomer
+from .greeter import Greeter
+from .visit import get_context
 
 
 @interface
@@ -33,21 +34,21 @@ class Greeting(Protocol):
 GreetingT = cast(Type[Greeting], Greeting)
 
 
-@implements(Greeting).when(qualified_by=DefaultCustomer)
+@implements(Greeting)
 @injectable
 @dataclass
 class DefaultGreeting:
     """The message given to a customer."""
 
-    customer: Customer = inject.me()
-    greeter: Greeter = inject.me(qualified_by=DefaultCustomer)
+    customer: Customer = get_context()
+    greeter: Greeter = inject.me(NotQualified())
     punctuation: str = MegaStoreConfig.PUNCTUATION
     salutation: str = "Hello"
 
     def __call__(self) -> str:
         """Give the text of the greeting."""
-        gn = self.greeter.name  # Shorten line
-        return f"{self.salutation}, my name is {gn}{self.punctuation}"
+        gn, p = self.greeter.name, self.punctuation  # Shorten line
+        return f"{self.salutation} {self.customer.name}, my name is {gn}{p}"
 
 
 @implements(Greeting).when(qualified_by=FrenchCustomer)
@@ -56,12 +57,12 @@ class DefaultGreeting:
 class FrenchGreeting:
     """The message given to a French customer."""
 
-    customer: Customer = inject.me()
+    customer: Customer = get_context()
     greeter: Greeter = inject.me(qualified_by=FrenchCustomer)
     punctuation: str = MegaStoreConfig.PUNCTUATION
     salutation: str = "Bonjour"
 
     def __call__(self) -> str:
         """Give the text of the greeting."""
-        gn = self.greeter.name  # Shorten line
-        return f"{self.salutation}, je m'appelle {gn}{self.punctuation}"
+        gn, p = self.greeter.name, self.punctuation  # Shorten line
+        return f"{self.salutation} {self.customer.name}, je m'appelle {gn}{p}"
