+++ src/antidote_book/interfaces/megastore_plugins/greeting.py
@@ -1,17 +1,32 @@
 """The message given by a Greeter to a Customer."""
 from dataclasses import dataclass
 
+from antidote import implements
 from antidote import inject
-from antidote import injectable
+from antidote import interface
 
-from ..config import MegaStoreConfig
-from ..customer import Customer
-from ..greeter import Greeter
+from .config import MegaStoreConfig
+from .customer import Customer
+from .greeter import Greeter
 
 
-@injectable
-@dataclass
+@interface
 class Greeting:
+    """A definition of a MegaStore greeting."""
+
+    customer: Customer
+    greeter: Greeter
+    punctuation: str
+    salutation: str
+
+    def __call__(self) -> str:
+        """Definition of the call method."""
+        return ""
+
+
+@implements(Greeting)
+@dataclass
+class DefaultGreeting(Greeting):
     """The message given to a customer."""
 
     customer: Customer = inject.me()
