+++ src/antidote_book/protocols/megastore_plugins/greeting.py
@@ -8,10 +8,9 @@ from antidote import implements
 from antidote import inject
 from antidote import interface
 
-from ..config import MegaStoreConfig
-from ..customer import Customer
-from ..greeter import Greeter
-from ..salutation import Salutation
+from .config import MegaStoreConfig
+from .customer import Customer
+from .greeter import Greeter
 
 
 @interface
@@ -21,7 +20,7 @@ class Greeting(Protocol):
     customer: Customer
     greeter: Greeter
     punctuation: str
-    salutation: Salutation
+    salutation: str
 
     def __call__(self) -> str:
         """Definition of the call method."""
@@ -39,7 +38,7 @@ class DefaultGreeting:
     customer: Customer = inject.me()
     greeter: Greeter = inject.me()
     punctuation: str = MegaStoreConfig.PUNCTUATION
-    salutation: Salutation = inject.me()
+    salutation: str = "Hello"
 
     def __call__(self) -> str:
         """Give the text of the greeting."""
