+++ src/antidote_book/scope/megastore_plugins/greeting.py
@@ -9,12 +9,14 @@ from antidote import inject
 from antidote import injectable
 from antidote import interface
 
-from ...megastore.predicates import NotQualified
-from ..config import MegaStoreConfig
-from ..customer import Customer
-from ..customer import FrenchCustomer
-from ..greeter import Greeter
-from ..visit import get_context
+from ..megastore.predicates import NotQualified
+from .config import MegaStoreConfig
+from .customer import ALL_CUSTOMERS
+from .customer import Customer
+from .customer import FrenchCustomer
+from .greeter import Greeter
+from .visit import Visit
+from .visit import VisitHandler
 
 
 @interface
@@ -40,15 +42,17 @@ GreetingT = cast(Type[Greeting], Greeting)
 class DefaultGreeting:
     """The message given to a customer."""
 
-    customer: Customer = get_context()
+    customer: Customer = inject.me()
     greeter: Greeter = inject.me(NotQualified())
     punctuation: str = MegaStoreConfig.PUNCTUATION
     salutation: str = "Hello"
+    visit: Visit = inject.me(source=VisitHandler)
 
     def __call__(self) -> str:
         """Give the text of the greeting."""
         gn, p = self.greeter.name, self.punctuation  # Shorten line
-        return f"{self.salutation} {self.customer.name}, my name is {gn}{p}"
+        customer = ALL_CUSTOMERS[self.visit.customer_id]
+        return f"{self.salutation} {customer.name}, my name is {gn}{p}"
 
 
 @implements(Greeting).when(qualified_by=FrenchCustomer)
@@ -57,12 +61,14 @@ class DefaultGreeting:
 class FrenchGreeting:
     """The message given to a French customer."""
 
-    customer: Customer = get_context()
+    customer: Customer = inject.me()
     greeter: Greeter = inject.me(qualified_by=FrenchCustomer)
     punctuation: str = MegaStoreConfig.PUNCTUATION
     salutation: str = "Bonjour"
+    visit: Visit = inject.me(source=VisitHandler)
 
     def __call__(self) -> str:
         """Give the text of the greeting."""
         gn, p = self.greeter.name, self.punctuation  # Shorten line
-        return f"{self.salutation} {self.customer.name}, je m'appelle {gn}{p}"
+        customer = ALL_CUSTOMERS[self.visit.customer_id]
+        return f"{self.salutation} {customer.name}, je m'appelle {gn}{p}"
