+++ src/antidote_book/vary/megastore_plugins/greeter.py
@@ -7,8 +7,8 @@ from typing import cast
 from antidote import implements
 from antidote import interface
 
-from ..customer import DefaultCustomer
-from ..customer import FrenchCustomer
+from .customer import DefaultCustomer
+from .customer import FrenchCustomer
 
 
 @interface
