+++ src/antidote_book/vary/megastore_plugins/greeting.py
@@ -6,12 +6,14 @@ from typing import cast
 
 from antidote import implements
 from antidote import inject
+from antidote import injectable
 from antidote import interface
 
-from ..config import MegaStoreConfig
-from ..customer import Customer
-from ..greeter import Greeter
-from ..salutation import Salutation
+from .config import MegaStoreConfig
+from .customer import Customer
+from .customer import DefaultCustomer
+from .customer import FrenchCustomer
+from .greeter import Greeter
 
 
 @interface
@@ -31,17 +33,35 @@ class Greeting(Protocol):
 GreetingT = cast(Type[Greeting], Greeting)
 
 
-@implements(Greeting)
+@implements(Greeting).when(qualified_by=DefaultCustomer)
+@injectable
 @dataclass
 class DefaultGreeting:
     """The message given to a customer."""
 
     customer: Customer = inject.me()
-    greeter: Greeter = inject.me()
+    greeter: Greeter = inject.me(qualified_by=DefaultCustomer)
     punctuation: str = MegaStoreConfig.PUNCTUATION
-    salutation: Salutation = inject.me()
+    salutation: str = "Hello"
 
     def __call__(self) -> str:
         """Give the text of the greeting."""
         gn = self.greeter.name  # Shorten line
         return f"{self.salutation}, my name is {gn}{self.punctuation}"
+
+
+@implements(Greeting).when(qualified_by=FrenchCustomer)
+@injectable
+@dataclass
+class FrenchGreeting:
+    """The message given to a French customer."""
+
+    customer: Customer = inject.me()
+    greeter: Greeter = inject.me(qualified_by=FrenchCustomer)
+    punctuation: str = MegaStoreConfig.PUNCTUATION
+    salutation: str = "Bonjour"
+
+    def __call__(self) -> str:
+        """Give the text of the greeting."""
+        gn = self.greeter.name  # Shorten line
+        return f"{self.salutation}, je m'appelle {gn}{self.punctuation}"
